Cleaning up user file paths - **Relative to compile_auth.py** - default to repository parent folder
Includes individual_keys and compile_auth.cfg; .gitignore has been updated to reflect these potential files.

A note on the config file: Either config or compile_auth.cfg will be searched in the repository, but only the latter in the parent folder.
Note that since the previous commit which introduced ../config was not yet pushed, this is inconsequential.